Move shared components back to components/
