+++ web/src/app/components/header/current-user-widget/current-user-widget.component.scss
