+++ web/src/app/components/header/current-user-widget/user-profile-popup/user-profile-popup.component.scss
