+++ web/src/app/components/header/current-user-widget/user-profile-popup/user-profile-popup.component.spec.ts
@@ -16,7 +16,7 @@
 
 import { AuthService } from 'app/services/auth/auth.service';
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { UserProfilePopupComponent } from 'app/shared/components/header/current-user-widget/user-profile-popup/user-profile-popup.component';
+import { UserProfilePopupComponent } from 'app/components/header/current-user-widget/user-profile-popup/user-profile-popup.component';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { MatDialogModule } from '@angular/material/dialog';
 import { Router } from '@angular/router';
