+++ web/src/app/components/user-avatar/user-avatar.component.scss
